requirements.txt
openai
torch
PyPDF2
ollama
streamlit
faiss-cpu
langchain